# contracts/katalog-api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Katalog.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/labels/search": {
      "get": {
        "tags": [
          "LabelsEndpoints"
        ],
        "operationId": "SearchLabels",
        "parameters": [
          {
            "name": "Q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/labels": {
      "get": {
        "tags": [
          "LabelsEndpoints"
        ],
        "operationId": "ListLabels",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabelSummaryResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "LabelsEndpoints"
        ],
        "operationId": "CreateLabel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLabelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelSummaryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/api/labels/{id}": {
      "get": {
        "tags": [
          "LabelsEndpoints"
        ],
        "operationId": "GetLabelDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "LabelsEndpoints"
        ],
        "operationId": "UpdateLabel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLabelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelSummaryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "delete": {
        "tags": [
          "LabelsEndpoints"
        ],
        "operationId": "DeleteLabel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/labels/{labelId}/artists": {
      "post": {
        "tags": [
          "LabelsEndpoints"
        ],
        "operationId": "AddArtistToLabel",
        "parameters": [
          {
            "name": "labelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddArtistToLabelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistSummaryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/labels/{labelId}/artists/{artistId}": {
      "delete": {
        "tags": [
          "LabelsEndpoints"
        ],
        "operationId": "RemoveArtistFromLabel",
        "parameters": [
          {
            "name": "labelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "artistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/search": {
      "get": {
        "tags": [
          "ArtistsEndpoints"
        ],
        "operationId": "SearchArtists",
        "parameters": [
          {
            "name": "Q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtistSearchResult"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/labels/{labelId}/releases": {
      "get": {
        "tags": [
          "ReleasesEndpoints"
        ],
        "operationId": "GetLabelReleases",
        "parameters": [
          {
            "name": "labelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddArtistToLabelRequest": {
        "required": [
          "spotifyArtistId"
        ],
        "type": "object",
        "properties": {
          "spotifyArtistId": {
            "type": "string"
          }
        }
      },
      "AlbumResponse": {
        "required": [
          "id",
          "spotifyId",
          "name",
          "albumType",
          "releaseDate",
          "releaseDatePrecision",
          "labelSpotify",
          "imageUrl",
          "externalUrl",
          "totalTracks",
          "artistNames"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "spotifyId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "albumType": {
            "type": "string"
          },
          "releaseDate": {
            "type": [
              "null",
              "string"
            ]
          },
          "releaseDatePrecision": {
            "type": "string"
          },
          "labelSpotify": {
            "type": [
              "null",
              "string"
            ]
          },
          "imageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "externalUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "totalTracks": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "artistNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ArtistSearchResult": {
        "required": [
          "id",
          "name",
          "imageUrl",
          "externalUrl",
          "genres",
          "popularity"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "imageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "externalUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "popularity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ArtistSummaryResponse": {
        "required": [
          "id",
          "spotifyId",
          "name",
          "imageUrl",
          "externalUrl",
          "genres",
          "popularity"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "spotifyId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "imageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "externalUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "genres": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "popularity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "CreateLabelRequest": {
        "required": [
          "name",
          "spotifyIds"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "spotifyIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "LabelDetailResponse": {
        "required": [
          "id",
          "spotifyIds",
          "name",
          "slug",
          "artistCount",
          "releaseCount",
          "createdAt",
          "updatedAt",
          "artists",
          "releases"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "spotifyIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "artistCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "releaseCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "artists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistSummaryResponse"
            }
          },
          "releases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlbumResponse"
            }
          }
        }
      },
      "LabelSearchAlbumResult": {
        "required": [
          "albumId",
          "name",
          "artists",
          "imageUrl",
          "releaseDate",
          "externalUrl"
        ],
        "type": "object",
        "properties": {
          "albumId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "artists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabelSearchArtistResult"
            }
          },
          "imageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "releaseDate": {
            "type": [
              "null",
              "string"
            ]
          },
          "externalUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "LabelSearchArtistResult": {
        "required": [
          "spotifyId",
          "name"
        ],
        "type": "object",
        "properties": {
          "spotifyId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "LabelSearchResponse": {
        "required": [
          "matchedLabelName",
          "albums"
        ],
        "type": "object",
        "properties": {
          "matchedLabelName": {
            "type": "string"
          },
          "albums": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabelSearchAlbumResult"
            }
          }
        }
      },
      "LabelSummaryResponse": {
        "required": [
          "id",
          "spotifyIds",
          "name",
          "slug",
          "artistCount",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "spotifyIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "artistCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateLabelRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "LabelsEndpoints"
    },
    {
      "name": "ArtistsEndpoints"
    },
    {
      "name": "ReleasesEndpoints"
    }
  ]
}